Made changes to starting a handstand
Inadvertently created cool concepts the player can enter handstand mid air and crouching in their air allows the player to glide. Also, very precise inputs allows the player to jump infinitely. Could be a cool mechanic to keep but I will make a fix for it in the next commit